+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 [Full changelog][unreleased]
 
+## Release 163 – 2025-01-23
+
+[Full changelog][163]
+
+- Add functionality to anonymise a user
+
 ## Release 162 – 2025-01-20
 
 [Full changelog][162]
@@ -1780,7 +1786,8 @@
 - Planned start and end dates are mandatory
 - Actual start and end dates must not be in the future
 
-[unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-162...HEAD
+[unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-163...HEAD
+[162]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-162...release-163
 [162]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-161...release-162
 [161]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-160...release-161
 [160]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-159...release-160